+++ src/app/routes/AUTH_MYPAGE_ROUTES.tsx
@@ -1,11 +1,11 @@
 import { PATH } from "@constants/path";
-import PersonalInfo from "@pages/MypageInfo";
+import MypageInfo from "@pages/MypageInfo";
 import Bookmark from "@pages/Bookmark";
 
 const AUTH_MYPAGE_ROUTES = [
   {
     path: PATH.information,
-    element: <PersonalInfo />,
+    element: <MypageInfo />,
   },
   {
     path: PATH.reservedList,
